+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## [1.0.0 - 02/02/2021]
 ### Added
 - Added new class to create and merge package and destructive salesforce files.
+- Created method to merge package XML files and destructive XML files
+- Created method to full merge package XML files or destructive XML files into one file XML file
+- Created methods to create Package XML file, Destructive Pre deploy XML File and Destructive Post deplot XML file
+- Create method to get the package XML content
+- Create method to validate the JSON Metadata file for create packages